fix (potentially partially) imports (in feature extraction test modules)
